requirements.txt
asgiref==3.6.0
certifi @ file:///work/ci_py311/certifi_1676822443922/work/certifi
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
psycopg2-binary==2.9.5
pytz==2023.2
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
